Move some definitions for setting to info.json
