+++ keyboards/hfdkb/keyboard_sw/k83/info.json
@@ -9,6 +9,16 @@
     },
     "processor": "WB32FQ95",
     "bootloader": "wb32-dfu",
+    "matrix_pins": {
+        "cols": ["C1","C2","C3","A0","A1","A2","A3","A4","A5","A6","A7","C4","C5","B0","B1","B2"],
+        "rows": ["B15", "C6", "C7", "C8", "C9", "A8"]
+    },
+    "diode_direction": "ROW2COL",
+    "encoder": {
+        "rotary": [
+            { "pin_a": "B14", "pin_b": "B13","resolution": 4 }
+        ]
+    },
     "layouts": {
         "LAYOUT": {
             "layout": [
